+++ zendriver/core/element.py
@@ -9,6 +9,7 @@ import pathlib
 import secrets
 import typing
 import urllib.parse
+import grapheme  # type: ignore
 
 from .. import cdp
 from . import util
@@ -690,19 +691,29 @@ class Element:
         """clears an input field"""
         return await self.apply('function (element) { element.value = "" } ')
 
-    async def send_keys(self, text: str):
+    async def send_keys(self, text: str, special_characters: bool = False):
         """
         send text to an input field, or any other html element.
 
         hint, if you ever get stuck where using py:meth:`~click`
         does not work, sending the keystroke \\n or \\r\\n or a spacebar work wonders!
 
+        when special_characters is True, it will use grapheme clusters to send the text:
+        if the character is in the printable ASCII range, it sends it using dispatch_key_event.
+        otherwise, it uses insertText, which handles special characters more robustly.
+
         :param text: text to send
+        :param special_characters: when True, uses grapheme clusters to send the text.
         :return: None
         """
         await self.apply("(elem) => elem.focus()")
-        for char in list(text):
-            await self._tab.send(cdp.input_.dispatch_key_event("char", text=char))
+        for cluster in grapheme.graphemes(text) if special_characters else text:
+            if all(32 <= ord(c) <= 126 for c in cluster):
+                await self._tab.send(
+                    cdp.input_.dispatch_key_event("char", text=cluster)
+                )
+            else:
+                await self._tab.send(cdp.input_.insert_text(cluster))
 
     async def send_file(self, *file_paths: PathLike):
         """